+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.16.0] - 2025-01-13
+
+### Added
+
+- `KPStandardIconButton` and `KPContainedIconButton` APIs introduced
+- Added Pause, Play, Save, Share icons
+
 ## [0.15.0] - 2025-01-10
 
 ### Changed